requirements.txt
aioredis==1.3.1
analytics-python==1.2.9
antlr4-python3-runtime==4.7.2
anyjson==0.3.3
asgiref==3.2.3
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==20.2.1
Automat==0.8.0
autopep8==1.5
awscli
awsebcli
backcall==0.1.0
bleach==3.1.0
boto==2.49.0
boto3==1.9.208
botocore
bson==0.5.8
cachetools==3.1.1
cement==2.8.2
certifi==2019.6.16
cffi==1.12.3
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
chart-studio==1.0.0
Click==7.0
colorama==0.3.9
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
Cython==0.29.16
daphne==2.4.1
dash==1.0.2
dash-core-components==1.0.0
dash-html-components==1.0.0
dash-renderer==1.0.0
dash-table==4.0.2
dataclasses==0.6
decorator==4.4.1
defusedxml==0.6.0
Django==2.2
django-activity-stream==0.8.0
django-admin-generator==2.2.0
django-anymail==6.1.0
django-appconf==1.0.3
django-autocomplete-light==3.3.5
django-bower==5.2.0
django-braces==1.14.0
django-contrib-comments==1.9.2
django-cryptography==0.3
django-debug-toolbar==2.2
django-extensions==2.2.1
django-hstore==1.4.2
django-jsonfield==1.2.0
django-load==1.0.0
django-notify-x==0.1.9
django-nvd3==0.9.7
django-pandas==0.6.1
django-pdb==0.6.2
django-picklefield==2.0
django-plotly-dash==1.0.1
django-polymorphic==2.1.2
django-ranged-response==0.2.0
django-silk==4.0.1
django-simple-captcha==0.5.12
django-storages==1.9.1
django-tables2==2.1.0
django-threadedcomments==1.2
django-utils2==2.5.4
djangorestframework==3.9.4
djangorestframework-api-key==1.4.1
djangorestframework-simplejwt==4.3.0
djongo==1.3.1
docutils==0.14
dpd-components==0.1.0
dropbox==9.4.0
entrypoints==0.3
Flask==1.1.1
Flask-Compress==1.4.0
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.14.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
googledrivedownloader==0.4
gprof2dot==2019.11.30
greenlet
gspread==3.6.0
gunicorn
hiredis==1.0.1
html5lib==1.0.1
httplib2==0.13.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.19
incremental==17.5.0
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.3
joblib==0.14.1
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
lxml
MarkupSafe==1.1.1
matplotlib==3.1.1
memory-profiler==0.57.0
mistune==0.8.4
more-itertools==7.2.0
msgpack==0.6.2
mysqlclient==1.4.4
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.1
npm==0.1.1
numexpr==2.7.1
numpy==1.16.4
oauth2client==4.1.3
oauthlib==3.1.0
onedrivesdk==1.1.8
optional-django==0.1.0
orca==1.5.3
packaging==19.1
pandas==1.0.1
pandocfilters==1.4.2
parse==1.14.0
parso==0.5.1
pathspec==0.5.5
patsy==0.5.1
pickleshare==0.7.5
Pillow==6.1.0
plotly==4.6.0
plotly-express==0.4.0
pluggy==0.12.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.1
psutil==5.6.7
psycopg2==2.8.3
public==2019.4.13
py==1.8.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycksum==0.4.3
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.6.1
PyHamcrest==2.0.0
PyJWT==1.7.1
pymongo==3.10.1
Pympler==0.8
pyOpenSSL==19.1.0
pyparsing==2.4.2
PyPDF2==1.26.0
PyQt5==5.13.0
PyQt5-sip==4.19.18
pyrsistent==0.15.5
pytest==5.1.0
pytest-django==3.5.1
python-dateutil==2.8.0
python-nvd3==0.14.2
python-slugify==1.1.4
python-utils==2.3.0
pytz==2019.1
pywinpty==0.5.5
PyYAML
pyzmq==18.1.0
qtconsole==4.5.5
query-string==2019.4.13
razorpay==1.2.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.2.0
retrying==1.3.3
rsa
s3transfer
scikit-learn==0.22.2.post1
scipy==1.3.0
seaborn==0.9.0
selenium==3.141.0
semantic-version==2.5.0
Send2Trash==1.5.0
service-identity==18.1.0
shell==1.0.1
six
sklearn==0.0
sqlparse==0.2.4
statsmodels==0.10.1
tables==3.6.1
tabulate==0.7.5
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toolz==0.10.0
tornado==6.0.3
traitlets==4.3.3
Twisted==19.10.0
txaio==20.1.1
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wincertstore==0.2
xhtml2pdf==0.2.4
xlrd==1.2.0
zipp==0.5.2
zope.interface==4.7.1
gevent